refactor(megatron): import module_utils from primus.core.utils
Point the migrated MLPerf patch modules at primus.core.utils.module_utils
instead of primus.modules.module_utils, which is being removed on main.
