+++ primus/backends/megatron/patches/parallelism/sdma_param_all_gather_patches.py
@@ -37,7 +37,7 @@ import os
 import torch
 
 from primus.core.patches import PatchContext, register_patch
-from primus.modules.module_utils import log_rank_0, warning_rank_0
+from primus.core.utils.module_utils import log_rank_0, warning_rank_0
 
 
 def _sdma_allgather_enabled(_ctx: PatchContext) -> bool:
