+++ primus/backends/megatron/patches/te_patches/bshd_layout_patches.py
@@ -30,7 +30,7 @@ import functools
 import os
 
 from primus.core.patches import PatchContext, register_patch
-from primus.modules.module_utils import log_rank_0, warning_rank_0
+from primus.core.utils.module_utils import log_rank_0, warning_rank_0
 
 
 def _bshd_enabled(_ctx: PatchContext) -> bool:
