+++ primus/backends/megatron/patches/turbo/fused_residual_norm_patches.py
@@ -26,7 +26,7 @@ gracefully otherwise, so it's safe to register unconditionally.
 import os
 
 from primus.core.patches import PatchContext, register_patch
-from primus.modules.module_utils import log_rank_0
+from primus.core.utils.module_utils import log_rank_0
 
 
 def _env_truthy(name: str) -> bool:
